requirements.txt
paho-mqtt
pandas
numpy
scikit-learn
joblib
requests